requirements.txt
Flask==2.3.3
numpy==1.24.3
faiss-cpu==1.7.4
requests==2.31.0
tensorflow==2.13.0
Werkzeug==2.3.7
gunicorn==21.2.0
Pillow==10.0.0
